scripts: restore: correct boot flag reset
If an update or a restore is a success, boot flags should be set in order to
control if the boot is a success or not. To enable this, "upgrade_available"
flag should be set to 1. Otherweise, boot status is not checked.

This patch enable it instead of clearing everything.

Signed-off-by: Pierre Marzin <pierre.marzin@iot.bzh>